Refactor cover-related functions: move to new covers helper file and clean up imports
