+++ components/shared/workCard/WorkCard.tsx
@@ -2,6 +2,7 @@ import Link from "next/link"
 import React from "react"
 
 import { WorkTeaserSearchPageFragment } from "@/lib/graphql/generated/fbi/graphql"
+import { getCoverUrls, getLowResCoverUrl } from "@/lib/helpers/covers"
 import { resolveUrl } from "@/lib/helpers/helper.routes"
 import { getIsbnsFromWork } from "@/lib/helpers/ids"
 import { useGetCoverCollection } from "@/lib/rest/cover-service-api/generated/cover-service"
@@ -11,7 +12,7 @@ import { useGetV1ProductsIdentifier } from "@/lib/rest/publizon-api/generated/pu
 import { Badge } from "../badge/Badge"
 import { CoverPicture } from "../coverPicture/CoverPicture"
 import WorkCardAvailabilityRow from "./WorkCardAvailabilityRow"
-import { displayCreators, getAllWorkPids, getCoverUrls, getLowResCoverUrl } from "./helper"
+import { displayCreators, getAllWorkPids } from "./helper"
 
 type WorkCardProps = {
   work: WorkTeaserSearchPageFragment
